Refactored WorkerHenlde constructor: Injected IOrchestratorService, renamed pingIpChecker to pingSender. Added AsyncMethodBreake structure. Improved dependency injection. Refactored logging for better debugging and maintenance.
