+++ TimeControl_Core/src/Modules/Break.Module.Core/Astractions/Irepository/IbreakRepositoryQeury.cs
@@ -3,7 +3,7 @@ namespace Modules.Break.Module.Core.Astractions.Irepository;
 
     public interface IbreakRepositoryQeury
     {
-        public Task<ResponseChecker<BrakeTime>> GetBreakByIdAsinc(int Id);
-        public Task<List<BrakeTime>> GetAllBreaksAsync();
+         Task<ResponseChecker<BrakeTime>> GetBreakByIdAsinc(int Id);
+         Task<List<BrakeTime>> GetAllBreaksAsync();
 
     }
