+++ TimeControl_Core/src/Modules/Break.Module.Core/BreakWorker/BackgroundService/WorkerHenlde.cs
@@ -6,25 +6,36 @@ namespace Modules.Break.Module.Core.BreakWorker.Command;
 
 public class WorkerHenlde : IWorkerHenlde
 {
-    
+
     private readonly ILogger<WorkerHenlde> logger;
-    private readonly IBreakTimeUpdateMediator breakeTimeMediator;
     private readonly IPingSender pingSender;
-    public WorkerHenlde(ILogger<WorkerHenlde> logger, IBreakTimeUpdateMediator breakeTimeMediator,IPingSender pingIpChecker)
-    => (this.breakeTimeMediator, this.logger, this.pingSender) = (breakeTimeMediator, logger, pingIpChecker);
+
+    private readonly IOrchestratorService brakeTimeService;
+    public WorkerHenlde(ILogger<WorkerHenlde> logger,IPingSender pingIpChecker, IOrchestratorService brakeTimeService)
+    => ( this.logger, this.pingSender, this.brakeTimeService) =(logger, pingIpChecker, brakeTimeService);
 
     public async Task AsyncMethodBreake()
     {
-      
+
         {
             try
             {
                 // get user  -->>>>   from db
 
-               // var PingResponseStatus = await pingSender.PingIp("192.168.100.3");
-              
-                  var PingResponseStatus = await pingSender.PingIp("192.168.1.94");
-                await breakeTimeMediator.UpdateBreakTimeAsync(1, PingResponseStatus);
+                // var PingResponseStatus = await pingSender.PingIp("192.168.100.3");
+
+                var PingResponseStatus = await pingSender.PingIp("192.168.1.94");
+
+                var workSchedule = new BrakeTimeDtoReqvest
+                {
+                    UserId = 1,
+                    StartTime = PingResponseStatus ? new List<DateTime>() : new List<DateTime> { DateTime.Now },
+                    EndTime = PingResponseStatus ? new List<DateTime> { DateTime.Now } : new List<DateTime>()
+                };
+
+                 await brakeTimeService.AddOrUpdateBrakeTime(workSchedule, PingResponseStatus);
+
+
             }
 
             catch (Exception ex)
