+++ TimeControl_Core/src/Modules/Break.Module.Core/DAL/Exstension/ServiceRegistration.cs
@@ -1,4 +1,8 @@
 
+
+
+using Break.Module.Core.BreakWorker.CommonServices.OrchestratorService;
+
 namespace Modules.Break.Module.Core.Exstension.DAL;
 
 public static class ServiceRegistration
@@ -36,9 +40,8 @@ public static class ServiceRegistration
 
         serviceDescriptors.AddHostedService<BreakWorkerCommand>();
         serviceDescriptors.AddScoped<IWorkerHenlde, WorkerHenlde>();
-        serviceDescriptors.AddScoped<IBreakTimeUpdateMediator, BreakTimeUpdateMediator>();
         serviceDescriptors.AddScoped<IServicesFactory, ServicesFactory>();
-        serviceDescriptors.AddScoped<IServicesFacade,ServicesFacade>();
+        serviceDescriptors.AddScoped<IServicesFacade,IServicesFacade>();
         serviceDescriptors.AddScoped<IOrchestratorService, OrchestratorService>();
         serviceDescriptors.AddScoped<IPersistenceService, PersistenceService>();
         serviceDescriptors.AddScoped<ITimeValidator, TimeValidator>();
